num_mods param should be 16 bits

+++ toxcore/group_moderation.c
@@ -31,7 +31,7 @@
  * Returns length of unpacked data on success.
  * Returns -1 on failure.
  */
-int mod_list_unpack(GC_Chat *chat, const uint8_t *data, uint32_t length, uint32_t num_mods)
+int mod_list_unpack(GC_Chat *chat, const uint8_t *data, uint32_t length, uint16_t num_mods)
 {
     if (length != num_mods * GC_MOD_LIST_ENTRY_SIZE) {
         return -1;
